feat: refactor, mobile pagination
